requirements.txt
python-telegram-bot==20.7
yt-dlp
instagrapi
moviepy==1.0.3